Enable other packages to vendor the simulator package
github.com/u-root/u-root has a little bit of TPM code, but it's hard to
test it because it vendors its dependencies, and go mod vendor deletes
files it doesn't think are relevant when resolving the vendored
dependency (i.e., it skips all the C files in ms-tpm-20-ref).

This change introduces placeholder .go files to allow other packages to vendor
this one, as suggested in https://github.com/golang/go/issues/26366.